Code reorg (#133)
