+++ crates/starknet-types-core/Cargo.toml
@@ -32,6 +32,7 @@ zeroize = { version = "1.8.1", default-features = false, optional = true }
 [features]
 default = ["std", "serde", "curve", "num-traits"]
 std = [
+  "alloc",
   "lambdaworks-math/std",
   "num-traits/std",
   "num-bigint/std",
@@ -52,7 +53,7 @@ papyrus-serialization = ["std"]
 zeroize = ["dep:zeroize"]
 
 [dev-dependencies]
-proptest = "1.5"
+proptest = { version = "1.5", default-features = false, features = ["alloc", "proptest-macro"] } 
 regex = "1.11"
 serde_test = "1"
 criterion = "0.5"
